@@ -1,12 +1,15 @@
 <?php
 
-Route::any('recipe.lists','RecipeController@lists');
-Route::any('recipe.info','RecipeController@info');
-Route::any('recipe.category','RecipeController@category');
-Route::any('recipe.search', 'RecipeController@search');
+use App\Http\Controllers\AuthController;
+use App\Http\Controllers\RecipeController;
 
-Route::post('login', 'AuthController@login');
-Route::post('logout', 'AuthController@logout');
-Route::post('register', 'AuthController@register');
+Route::any('recipe.lists', [RecipeController::class, 'lists']);
+Route::any('recipe.info', [RecipeController::class, 'info']);
+Route::any('recipe.category', [RecipeController::class, 'category']);
+Route::any('recipe.search', [RecipeController::class, 'search']);
 
-Route::resource('recipes', 'RecipeController');
+Route::post('login', [AuthController::class, 'login']);
+Route::post('logout', [AuthController::class, 'logout']);
+Route::post('register', [AuthController::class, 'register']);
+
+Route::resource('recipes', RecipeController::class);
